+++ experiments/ClimaEarth/run_dry_held_suarez.jl
@@ -40,8 +40,6 @@ include("components/atmosphere/climaatmos.jl")
 ## helpers for user-specified IO
 include("user_io/user_logging.jl")
 
-include("user_io/io_helpers.jl")
-
 #=
 ### Setup simulation parameters
 Here we follow ClimaCore's dry Held-Suarez `held_suarez_rhoe` example.
@@ -57,7 +55,7 @@ restart_t = Int(0)
 ## coupler simulation specific configuration
 Δt_cpl = Float64(400)
 t_end = "1000days"
-tspan = (Float64(0.0), Float64(time_to_seconds(t_end)))
+tspan = (Float64(0.0), Float64(Utilities.time_to_seconds(t_end)))
 start_date = "19790301"
 hourly_checkpoint = true
 
@@ -120,7 +118,7 @@ comms_ctx = Utilities.get_comms_context(Dict("device" => "auto"))
 ### I/O Directory Setup
 =#
 
-dir_paths = setup_output_dirs(output_dir = coupler_output_dir, comms_ctx = comms_ctx)
+dir_paths = Utilities.setup_output_dirs(output_dir = coupler_output_dir, comms_ctx = comms_ctx)
 @info(config_dict)
 
 #=